Mockup
A mockup showing basic function of the Zotero network graph. The
network graph depicts the relationships between "Related" Zotero items.
The network visualization appears in the top part of this diagram. In
the lower left is the current collection that is being depicted with
one item selected. In the lower right is the list of Related items for
the currently selected item from the lower left panel.